add jwt token

+++ graphql/resolvers/booking.js
@@ -1,10 +1,14 @@
 const Booking = require("../../models/booking");
 const Event = require("../../models/event");
-const { transformEvent, transformBooking, uid } = require("./merge");
+const { transformEvent, transformBooking } = require("./merge");
 
 module.exports = {
-  bookings: async () => {
+  bookings: async (args, req) => {
     try {
+      if (!req.isAuth) {
+        throw new Error("Unauthenticated!");
+      }
+
       const bookings = await Booking.find();
       return bookings.map(booking => {
         return transformBooking(booking);
@@ -13,11 +17,14 @@ module.exports = {
       throw err;
     }
   },
-  bookEvent: async args => {
+  bookEvent: async (args, req) => {
     try {
+      if (!req.isAuth) {
+        throw new Error("Unauthenticated!");
+      }
       const fetchedEvent = await Event.findOne({ _id: args.eventId });
       const booking = new Booking({
-        user: uid,
+        user: req.userId,
         event: fetchedEvent
       });
       const result = await booking.save();
@@ -26,8 +33,11 @@ module.exports = {
       throw err;
     }
   },
-  cancelBooking: async args => {
+  cancelBooking: async (args, req) => {
     try {
+      if (!req.isAuth) {
+        throw new Error("Unauthenticated!");
+      }
       const booking = await Booking.findById(args.bookingId).populate("event");
       const event = transformEvent(booking.event);
       await Booking.deleteOne({ _id: args.bookingId });
